+++ tests/ci/docker_images_check.py
@@ -6,10 +6,10 @@ import os
 import time
 import shutil
 from github import Github
-from report import create_test_html_report
 from s3_helper import S3Helper
 from pr_info import PRInfo
 from get_robot_token import get_best_robot_token, get_parameter_from_ssm
+from upload_result_helper import upload_results
 
 NAME = "Push to Dockerhub (actions)"
 
@@ -149,26 +149,6 @@ def process_test_results(s3_client, test_results, s3_path_prefix):
         processed_test_results.append((test_name, status))
     return overall_status, processed_test_results
 
-def upload_results(s3_client, pr_number, commit_sha, test_results):
-    s3_path_prefix = f"{pr_number}/{commit_sha}/" + NAME.lower().replace(' ', '_')
-
-    branch_url = "https://github.com/ClickHouse/ClickHouse/commits/master"
-    branch_name = "master"
-    if pr_number != 0:
-        branch_name = "PR #{}".format(pr_number)
-        branch_url = "https://github.com/ClickHouse/ClickHouse/pull/" + str(pr_number)
-    commit_url = f"https://github.com/ClickHouse/ClickHouse/commit/{commit_sha}"
-
-    task_url = f"https://github.com/ClickHouse/ClickHouse/actions/runs/{os.getenv('GITHUB_RUN_ID')}"
-
-    html_report = create_test_html_report(NAME, test_results, "https://hub.docker.com/u/clickhouse", task_url, branch_url, branch_name, commit_url)
-    with open('report.html', 'w') as f:
-        f.write(html_report)
-
-    url = s3_client.upload_test_report_to_s3('report.html', s3_path_prefix + ".html")
-    logging.info("Search result in url %s", url)
-    return url
-
 def get_commit(gh, commit_sha):
     repo = gh.get_repo(os.getenv("GITHUB_REPOSITORY", "ClickHouse/ClickHouse"))
     commit = repo.get_commit(commit_sha)
@@ -211,7 +191,6 @@ if __name__ == "__main__":
     else:
         description = "Nothing to update"
 
-
     if len(description) >= 140:
         description = description[:136] + "..."
 
@@ -220,7 +199,7 @@ if __name__ == "__main__":
     s3_path_prefix = str(pr_info.number) + "/" + pr_info.sha + "/" + NAME.lower().replace(' ', '_')
     status, test_results = process_test_results(s3_helper, images_processing_result, s3_path_prefix)
 
-    url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results)
+    url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, [], NAME)
 
     gh = Github(get_best_robot_token())
     commit = get_commit(gh, pr_info.sha)
