+++ tests/ci/functional_test_check.py
@@ -10,48 +10,12 @@ import sys
 
 from github import Github
 
-from report import create_test_html_report
 from s3_helper import S3Helper
 from get_robot_token import get_best_robot_token
 from pr_info import PRInfo
 from build_download_helper import download_all_deb_packages
+from upload_result_helper import upload_results
 
-def process_logs(s3_client, additional_logs, s3_path_prefix):
-    additional_urls = []
-    for log_path in additional_logs:
-        if log_path:
-            additional_urls.append(
-                s3_client.upload_test_report_to_s3(
-                    log_path,
-                    s3_path_prefix + "/" + os.path.basename(log_path)))
-
-    return additional_urls
-
-
-def upload_results(s3_client, pr_number, commit_sha, test_results, raw_log, additional_files, check_name):
-    additional_files = [raw_log] + additional_files
-    s3_path_prefix = f"{pr_number}/{commit_sha}/" + check_name.lower().replace(' ', '_').replace('(', '_').replace(')', '_').replace(',', '_')
-    additional_urls = process_logs(s3_client, additional_files, s3_path_prefix)
-
-    branch_url = "https://github.com/ClickHouse/ClickHouse/commits/master"
-    branch_name = "master"
-    if pr_number != 0:
-        branch_name = f"PR #{pr_number}"
-        branch_url = f"https://github.com/ClickHouse/ClickHouse/pull/{pr_number}"
-    commit_url = f"https://github.com/ClickHouse/ClickHouse/commit/{commit_sha}"
-
-    task_url = f"https://github.com/ClickHouse/ClickHouse/actions/runs/{os.getenv('GITHUB_RUN_ID')}"
-
-    raw_log_url = additional_urls[0]
-    additional_urls.pop(0)
-
-    html_report = create_test_html_report(check_name, test_results, raw_log_url, task_url, branch_url, branch_name, commit_url, additional_urls, True)
-    with open('report.html', 'w', encoding='utf-8') as f:
-        f.write(html_report)
-
-    url = s3_client.upload_test_report_to_s3('report.html', s3_path_prefix + ".html")
-    logging.info("Search result in url %s", url)
-    return url
 
 def get_commit(gh, commit_sha):
     repo = gh.get_repo(os.getenv("GITHUB_REPOSITORY", "ClickHouse/ClickHouse"))
@@ -221,7 +185,7 @@ if __name__ == "__main__":
 
     s3_helper = S3Helper('https://s3.amazonaws.com')
     state, description, test_results, additional_logs = process_results(result_path, server_log_path)
-    report_url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, run_log_path, additional_logs, check_name)
+    report_url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, [run_log_path] + additional_logs, check_name)
     print(f"::notice ::Report url: {report_url}")
     commit = get_commit(gh, pr_info.sha)
     commit.create_status(context=check_name, description=description, state=state, target_url=report_url)
