+++ tests/ci/pvs_check.py
@@ -8,26 +8,15 @@ import json
 import logging
 import sys
 from github import Github
-from report import create_test_html_report
 from s3_helper import S3Helper
 from pr_info import PRInfo
 from get_robot_token import get_best_robot_token, get_parameter_from_ssm
+from upload_result_helper import upload_results
 
 NAME = 'PVS Studio (actions)'
 LICENCE_NAME = 'Free license: ClickHouse, Yandex'
 HTML_REPORT_FOLDER = 'pvs-studio-html-report'
 TXT_REPORT_NAME = 'pvs-studio-task-report.txt'
-
-def process_logs(s3_client, additional_logs, s3_path_prefix):
-    additional_urls = []
-    for log_path in additional_logs:
-        if log_path:
-            additional_urls.append(
-                s3_client.upload_test_report_to_s3(
-                    log_path,
-                    s3_path_prefix + "/" + os.path.basename(log_path)))
-
-    return additional_urls
 
 def _process_txt_report(path):
     warnings = []
@@ -48,31 +37,6 @@ def get_commit(gh, commit_sha):
     repo = gh.get_repo(os.getenv("GITHUB_REPOSITORY", "ClickHouse/ClickHouse"))
     commit = repo.get_commit(commit_sha)
     return commit
-
-def upload_results(s3_client, pr_number, commit_sha, test_results, additional_files):
-    s3_path_prefix = str(pr_number) + "/" + commit_sha + "/" + NAME.lower().replace(' ', '_').replace('(', '_').replace(')', '_')
-    additional_urls = process_logs(s3_client, additional_files, s3_path_prefix)
-
-    branch_url = "https://github.com/ClickHouse/ClickHouse/commits/master"
-    branch_name = "master"
-    if pr_number != 0:
-        branch_name = "PR #{}".format(pr_number)
-        branch_url = "https://github.com/ClickHouse/ClickHouse/pull/" + str(pr_number)
-    commit_url = f"https://github.com/ClickHouse/ClickHouse/commit/{commit_sha}"
-
-    task_url = f"https://github.com/ClickHouse/ClickHouse/actions/runs/{os.getenv('GITHUB_RUN_ID')}"
-
-    raw_log_url = additional_urls[0]
-    additional_urls.pop(0)
-
-    html_report = create_test_html_report(NAME, test_results, raw_log_url, task_url, branch_url, branch_name, commit_url, additional_urls, False)
-    with open('report.html', 'w') as f:
-        f.write(html_report)
-
-    url = s3_client.upload_test_report_to_s3('report.html', s3_path_prefix + ".html")
-    logging.info("Search result in url %s", url)
-    return url
-
 
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO)
@@ -133,7 +97,7 @@ if __name__ == "__main__":
         test_results = [(index_html, "Look at the report"), ("Errors count not checked", "OK")]
         description = "Total errors {}".format(len(errors))
         additional_logs = [txt_report, os.path.join(temp_path, 'pvs-studio.log')]
-        report_url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, additional_logs)
+        report_url = upload_results(s3_helper, pr_info.number, pr_info.sha, test_results, additional_logs, NAME)
 
         print("::notice ::Report url: {}".format(report_url))
         commit = get_commit(gh, pr_info.sha)
