+++ tests/ci/split_build_smoke_check.py
@@ -9,17 +9,16 @@ import time
 
 from github import Github
 
-from report import create_test_html_report
 from s3_helper import S3Helper
 from get_robot_token import get_best_robot_token
 from pr_info import PRInfo
 from build_download_helper import download_shared_build
+from upload_result_helper import upload_results
 
 DOCKER_IMAGE = "clickhouse/split-build-smoke-test"
 DOWNLOAD_RETRIES_COUNT = 5
 RESULT_LOG_NAME = "run.log"
 CHECK_NAME = 'Split build smoke test (actions)'
-
 
 def process_result(result_folder, server_log_folder):
     status = "success"
@@ -51,43 +50,6 @@ def get_run_command(build_path, result_folder, server_log_folder, docker_image):
            f" --volume={result_folder}:/test_output" \
            f" {docker_image} >{result_folder}/{RESULT_LOG_NAME}"
 
-def process_logs(s3_client, additional_logs, s3_path_prefix):
-    additional_urls = []
-    for log_path in additional_logs:
-        if log_path:
-            additional_urls.append(
-                s3_client.upload_test_report_to_s3(
-                    log_path,
-                    s3_path_prefix + "/" + os.path.basename(log_path)))
-
-    return additional_urls
-
-def upload_results(s3_client, pr_number, commit_sha, test_results, additional_files, check_name):
-    s3_path_prefix = f"{pr_number}/{commit_sha}/" + check_name.lower().replace(' ', '_').replace('(', '_').replace(')', '_').replace(',', '_')
-    additional_urls = process_logs(s3_client, additional_files, s3_path_prefix)
-
-    branch_url = "https://github.com/ClickHouse/ClickHouse/commits/master"
-    branch_name = "master"
-    if pr_number != 0:
-        branch_name = f"PR #{pr_number}"
-        branch_url = f"https://github.com/ClickHouse/ClickHouse/pull/{pr_number}"
-    commit_url = f"https://github.com/ClickHouse/ClickHouse/commit/{commit_sha}"
-
-    task_url = f"https://github.com/ClickHouse/ClickHouse/actions/runs/{os.getenv('GITHUB_RUN_ID')}"
-
-    if additional_urls:
-        raw_log_url = additional_urls[0]
-        additional_urls.pop(0)
-    else:
-        raw_log_url = task_url
-
-    html_report = create_test_html_report(check_name, test_results, raw_log_url, task_url, branch_url, branch_name, commit_url, additional_urls, True)
-    with open('report.html', 'w', encoding='utf-8') as f:
-        f.write(html_report)
-
-    url = s3_client.upload_test_report_to_s3('report.html', s3_path_prefix + ".html")
-    logging.info("Search result in url %s", url)
-    return url
 
 def get_commit(gh, commit_sha):
     repo = gh.get_repo(os.getenv("GITHUB_REPOSITORY", "ClickHouse/ClickHouse"))
